travel_concierge/requirements.txt
google-adk>=1.5.0
python-dotenv>=1.0.0
beautifulsoup4>=4.9.0
requests>=2.28.0
lxml>=4.9.0
